[FLINK-25571] Update Elasticsearch Sink to use decomposed interfaces
